superpowered command handler
it now connects to a unix domain socket to fetch the fds to start the pinentry
loop. the next step (which is a WIP as you can see from the mess inside) is to pass
the pinentry_t to java land so the prompt can be displayed. this is basically done in
fill_struct()