check the version of app_opt, and if needed upgrade it
Since it takes a while to unpack the whole app_opt tree, it should only
happen when its needed. So this patch adds a VERSION file that contains the
Android versionCode, so the app knows when it needs to upgrade app_opt or
not. The old version is currently renamed by adding a timestamp to the name