resolve pinentry orientation changing issues
we're treating PinEntryActivity as a dialog that once dismissed
(either through back button, or some other event) CANNOT be reactivated.
The fact the pinentry dialog closed is propagated back to gpg-agent.

This mimicks the behavior of pinentry on the desktop, if you kill it,
click cancel, or click the X in your window manager, it dies and gpg-agent
must RELAUNCH pinentry if it wants to get the pin again.