gnupg-for-java: use POSIX FILE streams when creating GnuPGData for files
Previously, GnuPGData(File f) would load the entire contents of the file
into memory, then pass a memory gpgme_data_t to gpgme.  This is horrible when
working with large files or machines with limited memory.  This patch uses
the gpgme function for creating a gpgme_data_t from a POSIX FILE stream.