claim openpgp4fpr: URI scheme and launch dialog to receive from keyservers
GPG now registers to handle the 'openpgp4fpr:' URI scheme as a way to
direct OpenPGP 4 fingerprints to GPG for it to handle them.  Currently, it
prompts the user to download the key with that fingerprint from the PGP
keyservers.

openpgp4fpr:5E61C8780F86295CE17D86779F0FE587374BBE81

refs #471 closes #1594