add ContentProvider to read from private files, to sent via VIEW Intents
Currently, the data is decrypted to the private 'files' directory inside of
the app.  This is better than the SDcard, but it is still writing it to the
disk, which is less than desirable.  This is a temporary step towards
having data go from encrypted to viewable without touching the disk. Full
support for Streams needs to be implemented in gnupg-for-java (#1571) before that
is possible.

This commit adds the new PrivateFilesProvider, which uses GRANT_URI perms
to grant read perms based on the selected receiver of the VIEW intent.

refs #470 refs #1571 refs #1454 closes #1455 closes #1456