do NOT set noHistory on Activities that might pop up PinEntryActivity
When noHistory is set on an Activity, it is immediately killed when any
other Activity pops up.  In this case, PinEntryActivity will pop up when a
password is needed.  When that happens, the noHistory Activities get
killed, which breaks the onActivityResult() handling.

Instead, these Activities should just be sure to call finish() to remove
themselves from the history.