kludge to prevent FC when an Exception happens in onPostExecute()
This should be handled better, probably. It seems like this happens when
the Activity starts then stops to quickly, usually due to a problem with
the Activity that is supposed to handle the ACTION_VIEW after the crypto
action is complete.  This stuff should be replaced by a real Service with
a Notification.

refs #1482