load GnuPG libs needed by JNI via standard Android path
Now that the core libs are included in the standard Android way, we can
load the libs in the standard Android way. libgpg-error, libassuan, libgpgme,
and libpinentry are all needed by JNI stuff, so they are loaded in Java
space.  (libgcrypt and libkbsa are needed by things like gpg-agent, so they
don't need to be loaded into Java).

refs #1548