make project build depends work better so builds don't always get triggered
Currently, it seems that the dependencies, i.e. libgpg-error, get build
everytime no matter what.  This puts the depends on the installed .so file
so that should make it a bit more reliable.  Changing the installed .so
file will trigger a rebuild of that project and the dependent projects.
Not great, but better than before...

refs #1548