trigger Contacts sync based on BROADCAST_ACTION_KEYLIST_CHANGED
BROADCAST_ACTION_KEYLIST_CHANGED was already setup to be sent whenever the
GnuPG keylist changed in order to update the built-in key list viewer. This
commit hooks into the function that sends this broadcast and also requests
a sync with Contacts.

I stuck requestContactsSync() in sendKeylistChangedBroadcast() since it
seemed needlessly complicated to add a BroadcastReceiver to the
GpgApplication.

closes #1483