load global native libraries as soon as possible
Now that the libraries are included via ndk-build in the libs/ folder, and
unpacked from the APK by Android, that means the libs are guaranteed to be
in place before this app ever runs. So the libs can now be loaded as early
as possible by sticking them in a static{} block in GpgApplication.

refs #1548