search keyservers in a Loader and show results in KeyListFragment
This uses an AsyncTaskLoader to run the HTTP query to keep that off of the
UI thread.  It returns a custom whack job KeyserverResult so it can send
both the data and any errors that might have occured so that they can be
handled on the UI thread (i.e. Toast for now...)

refs #2808